wip: hopefully this will be completed or abandoned, and removed.
1. I added a `Commenter` class, that contains the logic needed to generate
a checkin/shelveset comment. I think its shape is ugly, but the extracted
function is nice. Maybe an extension method for `CheckinOptions` would be
better?

2. I started expunging `CheckinOptions` from TfsWorkspace. It was an ugly
hack to begin with, and shouldn't have been left around so long. It feels
like there is some refactoring possible that would make this class hold
together better.

3. I was just listening to some more-awesome-than-I guys talk about SOLID,
and I started to feel dumb about the IGitTfsRemote interface in regards to
the Interface Segregation Principle. So, I'm leaning towards, over time,
splitting it (and the other interfaces) up. GitTfsRemote may still be OK
as a single class (it probably needs to be split, too), but the interface
is too wide and handles too many unrelated things.