Highlight differences between <=VS2008 and >VS2008.
This refactoring focussed on highlighting what was the same
(or different) between VS2008 and newer TFS versions. The main
difference is that VS2008 does not sort changesets the way we
want on the server, so we need to query it in one lump.

This also changes the behavior of >VS2008 + !byLots, so that
everything gets fetched in one lump, so that we don't get only
the first 100 changesets.